+++ code/system.ui/react.common/src/test.ui/entry.Specs.mts
@@ -13,8 +13,8 @@ export const Specs = {
   'sys.ui.common.Text.Secret': () => import('../ui/Text.Secret/TextSecret.SPEC'),
   'sys.ui.common.Text.Syntax': () => import('../ui/Text.Syntax/TextSyntax.SPEC'),
   'sys.ui.common.Measure': () => import('../ui.tools/Measure/Measure.SPEC'),
-  'sys.ui.common.useFocus': () => import('../ui.hooks/Focus/useFocus.SPEC'),
-  'sys.ui.common.useSizeObserver': () => import('../ui.hooks/Size/useSizeObserver.SPEC'),
+  'sys.ui.common.useFocus': () => import('../ui/useFocus/useFocus.SPEC'),
+  'sys.ui.common.useSizeObserver': () => import('../ui/useSizeObserver/useSizeObserver.SPEC'),
 };
 
 export const DevSpecs = {
