+++ code/system.ui/react.common/src/types.mts
@@ -1,7 +1,6 @@
 export * from './ui/types.mjs';
 export * from './ui.dev/types.mjs';
 export * from './ui.tools/types.mjs';
-export * from './ui.hooks/types.mjs';
 
 /**
  * Common Types
