--- code/system.ui/react.common/src/ui.hooks/common.ts
@@ -1,1 +1,0 @@
-export * from '../common';
