--- code/system.ui/react.common/src/ui.hooks/index.ts
@@ -1,2 +1,0 @@
-export { useFocus } from './Focus';
-export { useSizeObserver } from './Size';
