--- code/system.ui/react.common/src/ui.hooks/types.mts
@@ -1,1 +1,0 @@
-export * from './Focus/types.mjs';
