+++ code/system.ui/react.common/src/ui/useSizeObserver/types.mts
@@ -1,0 +1,1 @@
+export {};
