+++ code/system.ui/react.common/src/ui/useSizeObserver/useSizeObserver.mts
